requirements.txt
absl-py==2.1.0
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.2.1
cloudpickle==3.1.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
databricks-sdk==0.60.0
datasets==4.0.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
docker==7.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.13.1
Flask==3.1.1
flatbuffers==24.3.7
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.3.0
gast==0.5.4
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
grpcio==1.62.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.10.0
hf-xet==1.1.5
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.34.1
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.30.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.1
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
libclang==18.1.1
Mako==1.3.10
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mlflow==3.1.4
mlflow-skinny==3.1.4
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
namex==0.0.7
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.3.1
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-cupti-cu12==12.3.101
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.3.107
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.0.12.1
nvidia-curand-cu12==10.3.4.107
nvidia-cusolver-cu12==11.5.4.101
nvidia-cusparse-cu12==12.2.0.103
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
omegaconf==2.3.0
opentelemetry-api==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opt-einsum==3.3.0
optree==0.10.0
optuna==4.4.0
overrides==7.7.0
packaging==24.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.3.0
pip==24.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.3.2
protobuf==4.25.3
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.17.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.22.3
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.33.2
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.2
sympy==1.13.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.36.0
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.2
tomli==2.2.1
torch==2.7.1+cpu
torchaudio==2.7.1+cpu
torchvision==0.22.1+cpu
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2024.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
wandb==0.21.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.16.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
